re-organises client code
