+++ client/src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import './App.css';
-import '../icon/IconLibrary';
+import './assets/styles/global.css';
+import './utils/IconLibrary';
 import { RouterProvider } from 'react-router-dom';
 import { Router } from '@remix-run/router';
