+++ client/src/components/error/ErrorHandler.tsx
@@ -1,8 +1,8 @@
 import React, { JSX } from 'react';
-import ErrorPage from '../page/ErrorPage';
+import ErrorPage from '../../pages/ErrorPage';
 import { useRouteError } from 'react-router-dom';
 import InternalServerError from './types/InternalServerError';
-import logError from '../../logger/Logger';
+import logError from '../../utils/Logger';
 
 type ErrorHandlerProps = {
   error: Error;
