+++ client/src/components/error/types/PageNotFoundError.ts
