+++ client/src/index.tsx
@@ -1,7 +1,7 @@
 import React, { StrictMode } from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './component/App';
-import router from './router/AppRouter';
+import App from './App';
+import router from './components/router/AppRouter';
 
 const root: ReactDOM.Root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
