+++ client/src/pages/ErrorPage.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ErrorPage from './ErrorPage';
 import { render, screen } from '@testing-library/react';
-import ERROR_MESSAGES from '../error/ErrorMessages';
+import ERROR_MESSAGES from '../components/error/ErrorMessages';
 
 describe('error page tests', () => {
   const expectedPageNotFoundMessage: string = ERROR_MESSAGES.PAGE_NOT_FOUND;
